extract collaboration user from collaboration class, and place it in a standalone class
